requirements.txt
flask==2.0.1
werkzeug==2.0.1
flask-cors==3.0.10
tensorflow==2.13.0
pillow==9.5.0
numpy==1.24.3